Move stress functions to ADV dir
