+++ mhkit/dolfyn/adv/turbulence.py
@@ -71,7 +71,7 @@ class ADVBinner(VelBinner):
 
         noise = ds.get('doppler_noise', [0, 0, 0])
         out['tke_vec'] = self.turbulent_kinetic_energy(ds['vel'], noise=noise)
-        out['stress'] = self.stresses(ds['vel'])
+        out['stress_vec'] = self.reynolds_stress(ds['vel'])
 
         out['psd'] = self.power_spectral_density(ds['vel'],
                                                  window=window,
@@ -86,6 +86,118 @@ class ADVBinner(VelBinner):
         out.attrs['n_fft_coh'] = self.n_fft_coh
 
         return out
+
+    def reynolds_stress(self, veldat, detrend=True):
+        """Calculate Reynolds stresses (cross-covariances of u,v,w in m^2/s^2)
+
+        Parameters
+        ----------
+        veldat : xr.DataArray
+            A velocity data array. The last dimension is assumed
+            to be time.
+        detrend : bool (default: True)
+            detrend the velocity data (True), or simply de-mean it
+            (False), prior to computing stress. Note: the psd routines
+            use detrend, so if you want to have the same amount of
+            variance here as there use ``detrend=True``.
+
+        Returns
+        -------
+        ds : xarray.DataArray
+
+        """
+        time = self.mean(veldat.time.values)
+        vel = veldat.values
+
+        out = np.empty(self._outshape(vel[:3].shape)[:-1],
+                       dtype=np.float32)
+
+        if detrend:
+            vel = self.detrend(vel)
+        else:
+            vel = self.demean(vel)
+
+        for idx, p in enumerate(self._cross_pairs):
+            out[idx] = np.nanmean(vel[p[0]] * vel[p[1]],
+                                  -1, dtype=np.float64
+                                  ).astype(np.float32)
+
+        da = xr.DataArray(out, name='stress_vec',
+                          dims=veldat.dims,
+                          attrs={'units': 'm^2/^2'})
+        da = da.rename({'dir': 'tau'})
+        da = da.assign_coords({'tau': ["upvp_", "upwp_", "vpwp_"],
+                               'time': time})
+        return da
+
+    def cross_spectral_density(self, veldat,
+                               freq_units='Hz',
+                               fs=None,
+                               window='hann',
+                               n_bin=None,
+                               n_fft_coh=None):
+        """Calculate the cross-spectral density of velocity components.
+
+        Parameters
+        ----------
+        veldat   : xarray.DataArray
+          The raw 3D velocity data.
+        freq_units : string
+          Frequency units of the returned spectra in either Hz or rad/s 
+          (`f` or :math:`\\omega`)
+        fs : float (optional)
+          The sample rate (default: from the binner).
+        window : string or array
+          Specify the window function.
+        n_bin : int (optional)
+          The bin-size (default: from the binner).
+        n_fft_coh : int (optional)
+          The fft size (default: n_fft_coh from the binner).
+
+        Returns
+        -------
+        csd : xarray.DataArray (3, M, N_FFT)
+          The first-dimension of the cross-spectrum is the three
+          different cross-spectra: 'uv', 'uw', 'vw'.
+
+        """
+        fs = self._parse_fs(fs)
+        n_fft = self._parse_nfft_coh(n_fft_coh)
+        time = self.mean(veldat.time.values)
+        veldat = veldat.values
+
+        out = np.empty(self._outshape_fft(veldat[:3].shape, n_fft=n_fft),
+                       dtype='complex')
+
+        # Create frequency vector, also checks whether using f or omega
+        coh_freq = self._fft_freq(units=freq_units, coh=True)
+        if 'rad' in freq_units:
+            fs = 2*np.pi*fs
+            freq_units = 'rad/s'
+            units = 'm^2/s/rad'
+            f_key = 'omega'
+        else:
+            freq_units = 'Hz'
+            units = 'm^2/s^2/Hz'
+            f_key = 'f'
+
+        for ip, ipair in enumerate(self._cross_pairs):
+            out[ip] = self._cpsd_ndarray(veldat[ipair[0]],
+                                         veldat[ipair[1]],
+                                         n_bin=n_bin,
+                                         n_fft=n_fft,
+                                         window=window)
+
+        da = xr.DataArray(out,
+                          name='csd',
+                          coords={'C': ['Cxy', 'Cxz', 'Cyz'],
+                                  'time': time,
+                                  f_key: coh_freq},
+                          dims=['C', 'time', f_key],
+                          attrs={'units': units, 'n_fft_coh': n_fft})
+        da[f_key].attrs['units'] = freq_units
+
+        return da
 
     def dissipation_rate_LT83(self, psd, U_mag, omega_range=[6.28, 12.57]):
         """
@@ -244,7 +356,7 @@ class ADVBinner(VelBinner):
         dat_avg : xarray.Dataset
           The bin-averaged adv dataset (calc'd from 'turbulence_statistics' or
           'bin_average'). The spectra (psd) and basic turbulence statistics 
-          ('tke_vec' and 'stress') must already be computed.
+          ('tke_vec' and 'stress_vec') must already be computed.
 
         Notes
         -----
